+++ typescript/sdk/src/contracts.ts
@@ -131,3 +131,28 @@ export async function filterOwnableContracts(
     (name, contract): contract is Ownable => isOwnableContracts[name],
   );
 }
+
+export function appFromAddressesMapHelper<F extends HyperlaneFactories>(
+  addressesMap: HyperlaneAddressesMap<any>,
+  factories: F,
+  multiProvider: MultiProvider,
+): {
+  contractsMap: HyperlaneContractsMap<F>;
+  multiProvider: MultiProvider;
+} {
+  // Attaches contracts for each chain for which we have a complete set of
+  // addresses
+  const contractsMap = attachContractsMap(addressesMap, factories);
+
+  // Filters out providers for chains for which we don't have a complete set
+  // of addresses
+  const intersection = multiProvider.intersect(Object.keys(contractsMap));
+
+  // Filters out contracts for chains for which we don't have a provider
+  const filteredContractsMap = pick(contractsMap, intersection.intersection);
+
+  return {
+    contractsMap: filteredContractsMap,
+    multiProvider: intersection.multiProvider,
+  };
+}
